server/pyproject.toml
[tool.poetry]
name = "knock-bank-api"
version = "0.1.0"
description = ""
authors = ["Davi Lucciola <davilucciola14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apiflask = "^2.0.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.2.0"
python-jose = "^3.3.0"
cpf-cnpj-validate = "^1.4"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
